requirements.txt
pytest>=8.0.0
pytest-html>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
playwright>=1.40.0
allure-pytest>=2.13.0
